requirements.txt
PyYAML
jinja2
docutils
babel
werkzeug
blinker
pygments
markdown